SurfaceWaterCalcTest.java total water remained tests converted to parametrized test.
Previously we had a couple of test methods looking similarly having input as profile and checking if it has correct water remained after rain. Now these methods replaced with just one @ParameterizedTest calculateWaterVolume with profile and expectedVolume as input - it uses surfaceProfileAndCalculatedWaterVolumeFixtures as a source input method. THis way we share same test method behavior but with different input and expected output.